requirements.txt
diff-match-patch==20121119
Django==1.10.3
django-import-export==1.0.0
django-sendfile==0.3.11
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.1
Pillow==4.2.1
psycopg2==2.6.2
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
